fix(connection): avoid calling close on write loop interrupt
Write loop interrupt is normally triggerred from outside of the loop and it is not necessary to close the connection
as the mechanism that stopped the loop has probably already done it.
Also lowered sleep period on connection close from 100ms to 5ms (to allow external routines cleanup before closing the channels).